==> code/00-preprocessing_delays.ipynb <==
import pandas as pd
import gtfs_kit as gk
import helper #homemade function file helper.py
import warnings
import numpy as np
import seaborn as sns
import dataframe_image as dfi
import json
import os
from datetime import datetime
from tqdm import tqdm

warnings.simplefilter(action='ignore', category=pd.errors.SettingWithCopyWarning)

lista_fermate = ['PISA CENTRALE','PISA SAN ROSSORE','S.GIULIANO TERME','RIGOLI','RIPAFRATTA','LUCCA','S. PIETRO A VICO',
                 'PONTE A MORIANO','DIECIMO - PESCAGLIA','BORGO A MOZZANO','BAGNI DI LUCCA','GHIVIZZANO - COREGLIA','FORNACI DI BARGA',
                 'BARGA GALLICANO','CASTELVECCHIO PASCOLI','FOSCIANDORA - CESERANA','CASTELNUOVO DI GARFAGNANA','VILLETTA S. ROMANO','POGGIO - CAREGGINE - VAGLI',
                 'CAMPORGIANO','PIAZZA AL SERCHIO','MINUCCIANO - PIEVE - CASOLA','EQUI TERME','MONZONE-M.DEI BANCHI-ISOLANO','GRAGNOLA',
                 'FIVIZZANO - GASSANO','FIVIZZANO ROMETTA - SOLIERA','AULLA LUNIGIANA']

mappa_location_codice = {
              'PISA CENTRALE':'S06500_1',
              'PISA SAN ROSSORE':'S06501_1',
              'S.GIULIANO TERME':'S06400_1',
              'RIGOLI':'S06401_1',
              'RIPAFRATTA':'S06402_1',
              'LUCCA':'S06404_1',
              'S. PIETRO A VICO':'S06313_1',
              'PONTE A MORIANO':'S06312_1',
              'DIECIMO - PESCAGLIA':'S06310_1',
              'BORGO A MOZZANO':'S06309_1',
              'BAGNI DI LUCCA':'S06308_1',
              'GHIVIZZANO - COREGLIA':'S06306_1',
              'FORNACI DI BARGA':'S06304_1',
              'BARGA GALLICANO':'S06303_1',
              'CASTELVECCHIO PASCOLI':'S06302_1',
              'FOSCIANDORA - CESERANA':'S06301_1',
              'CASTELNUOVO DI GARFAGNANA':'S06300_1',
              'VILLETTA S. ROMANO':'S06322_1',
              'POGGIO - CAREGGINE - VAGLI':'S06323_1',
              'CAMPORGIANO':'S06324_1',
              'PIAZZA AL SERCHIO':'S06325_1',
              'MINUCCIANO - PIEVE - CASOLA':'S06227_1',
              'EQUI TERME':'S06226_1',
              'MONZONE-M.DEI BANCHI-ISOLANO':  'S06225_1',
              'GRAGNOLA':'S06224_1',
              'FIVIZZANO - GASSANO':'S06223_1',
              'FIVIZZANO ROMETTA - SOLIERA':'S06222_1',
              'AULLA LUNIGIANA':'S06013_1'}

directory = '../../realtime_data'

delays_df =  pd.DataFrame(columns=['n','p','rp','a','ra','c','oaz','opz','op','oa','fr','pr','date'])

for filename in os.listdir(directory):
    file = os.path.join(directory, filename)
    if os.path.isfile(file):
        #df = pd.read_json('../../realtime_data/dati_01_01_2023.json')
        f = open(file)
        
        #Calcolo data in base al Nome del File
        data = json.load(f)
        giorno = f.name[25:35].split('_')
        date_to_add = giorno[2] + giorno[1] + giorno[0]
        #Load file in Dataframe
        df = pd.DataFrame(data['treni'])
        # only interested route
        df = df[df.p.isin(lista_fermate) & df.a.isin(lista_fermate)]
        df = df.reset_index(drop=True)
        ## Add date columns
        df['date'] = date_to_add
        ## Drop columns with all NAN
        df.dropna(axis=1, how='all')
        #Create one big dataset for delays
        delays_df = pd.concat([delays_df, df]).reset_index(drop=True)


delays_df = delays_df.rename(columns={'n': 'codice', 'p': 'punto_partenza','rp':'ritardo_partenza','a':'punto_arrivo',
                                     'ra':'ritardo_arrivo','c':'categoria'})
delays_df.dropna(axis=1, how='all')
delays_df = delays_df.reset_index(drop=True)

delays_df

for index, row in delays_df.iterrows():
    if(pd.notnull(delays_df['cn'][index])):
        value = delays_df.loc[index,'cn'].split(",")
        delays_df.loc[index,'codice'] = row['codice']+"*"+value[0]

#Delete train with wrong timestamps 
delays_df['op'] = delays_df['op'].astype(str)
delays_df['oa'] = delays_df['oa'].astype(str)

delays_df['oaz_len'] = delays_df.oaz.str.len()
delays_df['opz_len'] = delays_df.opz.str.len()
delays_df['op_len'] = delays_df.op.str.len()
delays_df['oa_len'] = delays_df.oa.str.len()

delays_df['op'] = delays_df['op'].astype('int')
delays_df['oa'] = delays_df['oa'].astype('int')

delays_df = delays_df[delays_df.oaz_len == 10]
delays_df = delays_df[delays_df.opz_len == 10]
delays_df = delays_df[delays_df.op_len == 10]
delays_df = delays_df[delays_df.oa_len == 10]

delays_df = delays_df.drop(['oaz_len', 'opz_len','op_len','oa_len'], axis=1)

delays_df.shape 

delays_df

delays_df.to_csv('../processed_files/delays_preprocessing_general_dataset.csv')


# We now expand the dataset by having trip_id, date, station
delays_expanded = pd.DataFrame(columns=['codice','data_giorno','codice_stazione','opt','oat'])
max_val = delays_df.index.max()
for index, row  in delays_df.iterrows():
    if(index % 1000 == 0):
        print("row " + str(index) + "/" +str(max_val))
    json_to_put_in_table = str(delays_df.loc[index, 'fr']).replace("'","\"")
    if(json_to_put_in_table != "nan"):
        temp_dictionary = json.loads(json_to_put_in_table)
        df = pd.DataFrame(temp_dictionary)
        df['codice'] = row['codice']
        df['data_giorno'] = row['date']
        df.oa = df.oa.apply(lambda x: datetime.fromtimestamp(x).time())
        df.op = df.op.apply(lambda x: datetime.fromtimestamp(x).time())
        delays_expanded = pd.concat([delays_expanded, df]).reset_index(drop=True)
    else:
        
        row_partenza = {'codice': row['codice'], 'data_giorno':row['date'],'codice_stazione':'', 'n':row['punto_partenza'], 
                        'ra': 'N', 'rp' : row['ritardo_partenza'],'oa': 0,'op': datetime.fromtimestamp(row['op']).time() }
        
        row_arrivo = {'codice': row['codice'], 'data_giorno':row['date'], 'codice_stazione':'','n':row['punto_arrivo'], 
                        'ra': row['ritardo_arrivo'] , 'rp' : 'N','oa': datetime.fromtimestamp(row['oa']).time(),'op': 0 }
        
        #row_partenza = {'codice': row['codice'], 'data_giorno':row['date'],'codice_stazione':'', 'n':row['punto_partenza'], 
        #                'ra': 'N', 'rp' : row['ritardo_partenza']}
        
        #row_arrivo = {'codice': row['codice'], 'data_giorno':row['date'], 'codice_stazione':'','n':row['punto_arrivo'], 
        #                'ra': row['ritardo_arrivo'] , 'rp' : 'N' }
        
        delays_expanded = pd.concat([delays_expanded, pd.DataFrame([row_partenza])], ignore_index=True)
        delays_expanded = pd.concat([delays_expanded, pd.DataFrame([row_arrivo])], ignore_index=True)

delays_expanded = delays_expanded.rename(columns={'n': 'stazione', 'ra': 'ritardo_arrivo',
                                                  'rp':'ritardo_partenza','oa':'orario_arrivo',
                                                  'op':'orario_partenza'})
 
delays_expanded["codice_stazione"] = delays_expanded["stazione"].apply(lambda x: mappa_location_codice.get(x))


delays_expanded = delays_expanded.drop(['stazione'], axis=1)

delays_expanded.to_csv('../processed_files/delays_preprocessing_expanded.csv')

delays_expanded
