==> code/00-preprocessing_weather.ipynb <==
# Import

import pandas as pd
import warnings

warnings.simplefilter(action='ignore', category=pd.errors.SettingWithCopyWarning)

# Load

locations = pd.read_csv('../../weather/locations.csv')
measurments = pd.read_csv('../../weather/mesurments.csv')

# Code

mappa_location_codice = {
              0:'S06500_1',
              1:'S06501_1',
              2:'S06400_1',
              3:'S06401_1',
              4:'S06402_1',
              5:'S06404_1',
              6:'S06313_1',
              7:'S06312_1',
              8:'S06310_1',
              9:'S06309_1',
              10:'S06308_1',
              11:'S06306_1',
              12:'S06304_1',
              13:'S06303_1',
              14:'S06302_1',
              15:'S06301_1',
              16:'S06300_1',
              17:'S06322_1',
              18:'S06323_1',
              19:'S06324_1',
              20:'S06325_1',
              21:'S06227_1',
              22:'S06226_1',
              23:'S06225_1',
              24:'S06224_1',
              25:'S06223_1',
              26:'S06222_1',
              27:'S06013_1'}

locations["codice_stazione"] = locations["location_id"].apply(lambda x: mappa_location_codice.get(x))

measurments.head()

### Clean up time in measurment table

measurments = measurments.assign(data_giorno=None)
measurments = measurments.assign(orario=None)


for index, row  in measurments.iterrows():
    if(index % 10000 == 0):
        print("row " + str(index) + "/" +str(measurments.index.max()))
    date,time = measurments.loc[index,"time"].split("T")
    measurments.loc[index,'orario'] = time
    measurments.loc[index,'data_giorno'] = date.replace("-","")

# Creo una tabella unica
measurments["codice_stazione"] = measurments["location_id"].apply(lambda x: mappa_location_codice.get(x))

measurments = measurments.rename(columns={'temperature_2m (°C)': 'temperature',
                                          'apparent_temperature (°C)' : 'app_temp',
                                          'relative_humidity_2m (%)' : 'humidity',
                                          'precipitation (mm)' : 'precipitation',
                                          'rain (mm)' : 'rain', 
                                          'snowfall (cm)':'snow_fall',
                                          'snow_depth (m)':'snow_depth',
                                          'weather_code (wmo code)' : 'wmo_code',
                                          'wind_speed_10m (km/h)':'wind_speed_10m',
                                          'wind_speed_100m (km/h)':'wind_speed_100m',
                                          'wind_direction_10m (°)':'wind_dir_10m',
                                          'wind_direction_100m (°)':'wind_dir_100m',
                                          'wind_gusts_10m (km/h)' : 'wind_gusts',
                                          'soil_temperature_0_to_7cm (°C)':'soil_temperature'
                                         })

measurments['data_giorno'] = measurments['data_giorno'].astype(str)

measurments.to_csv('../processed_files/measurments_preprocessing.csv')

measurments
