==> code/01-preprocessing.ipynb <==
# Import

import pandas as pd
import gtfs_kit as gk
import helper #homemade function file helper.py
import warnings
import numpy as np
from tqdm import tqdm

warnings.simplefilter(action='ignore', category=pd.errors.SettingWithCopyWarning)

# Main Dataset Creation

# Load GTFS Static
feed = gk.feed.read_feed('../processed_files/preprocessing_gtfs_static.zip',dist_units="km")
routes = gk.routes.get_routes(feed)
stop_times = feed.get_stop_times()
trips = feed.get_trips()
stops = feed.get_stops()
shapes = feed.shapes
calendar_dates = feed.calendar_dates
calendar = feed.calendar

# Load Delays
delays = pd.read_csv('../processed_files/delays_preprocessing_expanded.csv',index_col=[0])
# Load Weather
weather = pd.read_csv('../processed_files/measurments_preprocessing.csv',index_col=[0])



main_dataset = stop_times.copy()
main_dataset = pd.merge(main_dataset, trips, on=['trip_id','trip_id']).reset_index(drop=True)
main_dataset = pd.merge(main_dataset, calendar_dates, on=['service_id','service_id']).reset_index(drop=True)
main_dataset = pd.merge(main_dataset, stops, on=['stop_id','stop_id']).reset_index(drop=True)
main_dataset = main_dataset.drop(['route_id', 'service_id','trip_headsign', 'shape_id',
                                 'geometry','stop_name'], axis=1)

### Keep only rows where date are compatible with delays and weather
main_dataset.shape

main_dataset.sort_values('trip_short_name',ascending=False)

main_dataset = main_dataset.rename(columns={'trip_short_name': 'codice','stop_id':'codice_stazione','date':'data_giorno'})

main_dataset = main_dataset.loc[(main_dataset["data_giorno"] >= "20200110") & (main_dataset["data_giorno"] <= "20231031")]

main_dataset.shape

#Set Join columns to same dtype
delays["codice"] = delays["codice"].astype(str)
delays["data_giorno"] = delays["data_giorno"].astype(str)

#Join on Delays
main_dataset = pd.merge(main_dataset, delays, on=['codice', 'data_giorno','codice_stazione'])

main_dataset.shape

#Calcolo orario arrotondato per join con weather.
main_dataset['orario'] = main_dataset.loc[:, 'arrival_time']
max_index_for_display = main_dataset.index.max()
for index, row  in tqdm(main_dataset.iterrows()):
    orario_to_split = main_dataset.loc[index,"orario"].split(":")
    if orario_to_split[1] >= "30":
        orario_to_split[0]= str( int(orario_to_split[0]) + 1 )
    main_dataset.loc[index,"orario"] = orario_to_split[0] + ":00"

#Join on Weather
weather['data_giorno'] = weather['data_giorno'].astype(str)
main_dataset = pd.merge(main_dataset, weather, on=['data_giorno','orario','codice_stazione'])
main_dataset.shape



#Sort Dataset
main_dataset = main_dataset.sort_values(['trip_id','data_giorno','stop_sequence'],ascending=True)
main_dataset = main_dataset.reset_index(drop=True)

# We create an empty column for time different, speed, distance 
main_dataset = main_dataset.assign(temperature_linea=None)
main_dataset = main_dataset.assign(app_temp_linea=None)
main_dataset = main_dataset.assign(rain_linea=None)
main_dataset = main_dataset.assign(snow_fall_linea=None)
main_dataset = main_dataset.assign(snow_depth_linea=None)
main_dataset = main_dataset.assign(humidity_linea=None)
main_dataset = main_dataset.assign(precipitation_linea=None)
main_dataset = main_dataset.assign(wmo_code_linea=None)
main_dataset = main_dataset.assign(wind_speed_10m_linea=None)
main_dataset = main_dataset.assign(wind_speed_100m_linea=None)
main_dataset = main_dataset.assign(wind_dir_10m_linea=None)
main_dataset = main_dataset.assign(wind_dir_100m_linea=None)
main_dataset = main_dataset.assign(wind_gusts_linea=None)
main_dataset = main_dataset.assign(soil_temperature_linea=None)

max_index_for_display = main_dataset.index.max()

# Calculating Distance, Time and avg Speed between last stop and current stop
for index, row  in tqdm(main_dataset.iterrows()):
    if row['stop_sequence'] == 1:
        main_dataset['temperature_linea'][index] = row['temperature']
        main_dataset['app_temp_linea'][index] = row['app_temp']
        main_dataset['rain_linea'][index] = row['rain']
        main_dataset['snow_fall_linea'][index] = row['snow_fall']
        main_dataset['snow_depth_linea'][index] = row['snow_depth']
        main_dataset['humidity_linea'][index] = row['humidity']
        main_dataset['precipitation_linea'][index] = row['precipitation']
        main_dataset['wind_speed_10m_linea'][index] = row['wind_speed_10m']
        main_dataset['wind_speed_100m_linea'][index] = row['wind_speed_100m']
        main_dataset['wind_dir_10m_linea'][index] = row['wind_dir_10m']
        main_dataset['wind_dir_100m_linea'][index] = row['wind_dir_100m']
        main_dataset['wmo_code_linea'][index] = row['wmo_code']
        main_dataset['wind_gusts_linea'][index] = row['wind_gusts']
        main_dataset['soil_temperature_linea'][index] = row['soil_temperature']
    else:
        main_dataset['temperature_linea'][index] = ((row['temperature'] + main_dataset.loc[index - 1,'temperature']) / 2) 
        main_dataset['app_temp_linea'][index] = ((row['app_temp'] + main_dataset.loc[index - 1,'app_temp']) / 2) 
        main_dataset['rain_linea'][index] = ((row['rain'] + main_dataset.loc[index - 1,'rain']) / 2) 
        main_dataset['snow_fall_linea'][index] = ((row['snow_fall'] + main_dataset.loc[index - 1,'snow_fall']) / 2) 
        main_dataset['snow_depth_linea'][index] = ((row['snow_depth'] + main_dataset.loc[index - 1,'snow_depth']) / 2) 
        main_dataset['precipitation_linea'][index] = ((row['precipitation'] + main_dataset.loc[index - 1,'precipitation']) / 2) 
        main_dataset['humidity_linea'][index] = ((row['humidity'] + main_dataset.loc[index - 1,'humidity']) / 2)
        main_dataset['wind_speed_10m_linea'][index] = ((row['wind_speed_10m'] + main_dataset.loc[index - 1,'wind_speed_10m']) / 2) 
        main_dataset['wind_speed_100m_linea'][index] =((row['wind_speed_100m'] + main_dataset.loc[index - 1,'wind_speed_100m']) / 2)  
        main_dataset['wind_dir_10m_linea'][index] = ((row['wind_dir_10m'] + main_dataset.loc[index - 1,'wind_dir_10m']) / 2) 
        main_dataset['wind_dir_100m_linea'][index] = ((row['wind_dir_100m'] + main_dataset.loc[index - 1,'wind_dir_100m']) / 2)
        main_dataset['wmo_code_linea'][index] = ((row['wmo_code'] + main_dataset.loc[index - 1,'wmo_code']) / 2) 
        main_dataset['wind_gusts_linea'][index] = ((row['wind_gusts'] + main_dataset.loc[index - 1,'wind_gusts']) / 2) 
        main_dataset['soil_temperature_linea'][index] = ((row['soil_temperature'] + main_dataset.loc[index - 1,'soil_temperature']) / 2)

main_dataset

main_dataset=main_dataset.dropna().reset_index(drop=True)


main_dataset

# Drop Duplicates
row_to_delete =[]  
# Change orario realtime arrivo stazione 1 (non c'è to = arrivo_schedulato)
for index, row  in tqdm(main_dataset.iterrows()):
    # X nessun ritardo, N significa che non riparte perché era l'ultima destinazione oppure quella di partenza
    if((row['ritardo_arrivo'] == 'X') |( row['ritardo_arrivo'] == 'N')):
        main_dataset.loc[index,"ritardo_arrivo"] = '0'
    if((row['ritardo_partenza'] == 'X' )| (row['ritardo_partenza'] == 'N')):
        main_dataset.loc[index,"ritardo_partenza"] = '0'
    if((row['ritardo_arrivo'] == 'S') | (row['ritardo_arrivo'] == 'n.d.')):
        row_to_delete.append(index)
    if((row['ritardo_partenza'] == 'S') | (row['ritardo_partenza'] == 'n.d.')):
        row_to_delete.append(index)

row_to_delete = list(dict.fromkeys(row_to_delete))
main_dataset = main_dataset[~main_dataset.index.isin(row_to_delete)]
main_dataset.reset_index(drop=True)

main_dataset["ritardo_partenza"] = main_dataset["ritardo_partenza"].astype(int)
main_dataset["ritardo_arrivo"] = main_dataset["ritardo_arrivo"].astype(int)
max_index_for_display = main_dataset.index.max()

for index, row  in tqdm(main_dataset.iterrows()):
    main_dataset.loc[index,'arrival_time'] = row['arrival_time'].replace(':','')
    main_dataset.loc[index,'departure_time'] = row['departure_time'].replace(':','')
    main_dataset.loc[index,'stop_code'] = row['stop_code'].replace('S','')
    main_dataset.loc[index,'time_diff'] = round(row['time_diff'], 2)
    main_dataset.loc[index,'dist_diff'] = round(row['dist_diff'], 2)
    main_dataset.loc[index,'speed'] = round(row['speed'], 2)
    
    if row['ritardo_arrivo'] < 0:
        main_dataset.loc[index,'ritardo_arrivo'] = 0
    if row['ritardo_partenza'] < 0:
        main_dataset.loc[index,'ritardo_partenza'] = 0


main_dataset

main_dataset.to_csv("../processed_files/main_dataset.csv")

#main_dataset = pd.read_csv("../processed_files/main_dataset.csv", index_col=[0])
main_dataset = main_dataset.reset_index(drop=True)

df_linea =  pd.DataFrame(columns=['trip_id','arrival_time','departure_time','codice_stazione_p',
                                  'codice_stazione_a','indice_tratta','shape_dist_traveled',
                                  'time_diff','speed','dist_diff','codice','direction_id','data_giorno',
                                  'stop_code_p','stop_code_a','ritardo_partenza','ritardo_arrivo',
                                  'temperature_linea','app_temp_linea','rain_linea','snow_fall_linea','snow_depth_linea','humidity_linea',
                                  'precipitation_linea','wmo_code_linea','wind_speed_10m_linea','wind_speed_100m_linea',
                                  'wind_dir_10m_linea','wind_dir_100m_linea','wind_gusts_linea','soil_temperature_linea'])


# Calculating Distance, Time and avg Speed between last stop and current stop
for index, row  in tqdm(main_dataset.iterrows()):
    next_row = main_dataset.iloc[[index+1]]
    check1 = (next_row['trip_id'] == row['trip_id']).bool()
    check2 = (next_row['data_giorno'] == row['data_giorno']).bool()
    if(check1 & check2):
        row_to_concat = {'trip_id': row['trip_id'],
                         'departure_time':row['departure_time'],
                         'arrival_time':next_row['arrival_time'].values,
                         'codice_stazione_p':row['codice_stazione'],
                         'codice_stazione_a': next_row['codice_stazione'].values[0],
                         'indice_tratta':row['stop_sequence'],
                         'shape_dist_traveled':next_row['shape_dist_traveled'].values,
                         'time_diff':next_row['time_diff'].values,
                         'speed':next_row['speed'].values,
                         'dist_diff':next_row['dist_diff'].values,
                         'codice':row['codice'],
                         'direction_id':row['direction_id'],
                         'data_giorno':row['data_giorno'],
                         'stop_code_p':row['stop_code'],
                         'stop_code_a':next_row['stop_code'].values[0],
                         'ritardo_partenza':row['ritardo_partenza'],
                         'ritardo_arrivo':next_row['ritardo_arrivo'].values,
                         'temperature_linea':next_row['temperature_linea'].values,
                         'app_temp_linea':next_row['app_temp_linea'].values,
                         'rain_linea':next_row['rain_linea'].values,
                         'snow_fall_linea':next_row['snow_fall_linea'].values,
                         'snow_depth_linea':next_row['snow_depth_linea'].values,
                         'humidity_linea':next_row['humidity_linea'].values,
                         'precipitation_linea':next_row['precipitation_linea'].values,
                         'wmo_code_linea':next_row['wmo_code_linea'].values,
                         'wind_speed_10m_linea':next_row['wind_speed_10m_linea'].values,
                         'wind_speed_100m_linea':next_row['wind_speed_100m_linea'].values,
                         'wind_dir_10m_linea':next_row['wind_dir_10m_linea'].values,
                         'wind_dir_100m_linea':next_row['wind_dir_100m_linea'].values,
                         'wind_gusts_linea': next_row['wind_gusts_linea'].values,
                         'soil_temperature_linea':next_row['soil_temperature_linea'].values
                    }    
        df_linea = pd.concat([df_linea, pd.DataFrame([row_to_concat])], ignore_index=True)


df_linea['arrival_time'] = df_linea['arrival_time'].astype(int)
#df_linea['codice_stazione_a'] = 
df_linea['shape_dist_traveled'] = df_linea['shape_dist_traveled'].astype(float)
df_linea['time_diff'] = df_linea['time_diff'].astype(float)
df_linea['speed'] = df_linea['speed'].astype(float)
df_linea['dist_diff'] = df_linea['dist_diff'].astype(float)
#df_linea['stop_code_a'] = df_linea['stop_code_a'].transform(lambda x: ','.join(x))
df_linea['ritardo_arrivo'] = df_linea['ritardo_arrivo'].astype(int)
df_linea['temperature_linea'] = df_linea['temperature_linea'].astype(float)
df_linea['app_temp_linea'] = df_linea['app_temp_linea'].astype(float)
df_linea['rain_linea'] = df_linea['rain_linea'].astype(float)
df_linea['snow_fall_linea'] = df_linea['snow_fall_linea'].astype(float)
df_linea['snow_depth_linea'] = df_linea['snow_depth_linea'].astype(float)
df_linea['humidity_linea'] = df_linea['humidity_linea'].astype(float)
df_linea['precipitation_linea'] = df_linea['precipitation_linea'].astype(float)
df_linea['wmo_code_linea'] = df_linea['wmo_code_linea'].astype(float)
df_linea['wind_speed_10m_linea'] = df_linea['wind_speed_10m_linea'].astype(float)
df_linea['wind_speed_100m_linea'] = df_linea['wind_speed_100m_linea'].astype(float)
df_linea['wind_dir_10m_linea'] = df_linea['wind_dir_10m_linea'].astype(float)
df_linea['wind_dir_100m_linea'] = df_linea['wind_dir_100m_linea'].astype(float)
df_linea['wind_gusts_linea'] = df_linea['wind_gusts_linea'].astype(float)
df_linea['soil_temperature_linea'] = df_linea['soil_temperature_linea'].astype(float)

df_linea = df_linea.reset_index(drop=True)
df_linea.to_csv("../processed_files/df_linea.csv")
