==> code/06-regressor.ipynb <==
# Import

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import joblib
import helper

from sklearn.model_selection import train_test_split
from sklearn.linear_model import LinearRegression
from sklearn.tree import DecisionTreeRegressor
from sklearn.ensemble import RandomForestRegressor
from sklearn.ensemble import ExtraTreesRegressor
from sklearn.ensemble import GradientBoostingRegressor
from sklearn.ensemble import BaggingRegressor
from sklearn.ensemble import AdaBoostRegressor

# Load Dataset

df_regression = pd.read_csv("../processed_files/df_for_ml.csv", index_col=[0])

df_regression.info()

# Drop column on wich i don't want to train
df_regression.drop(columns=["trip_id",
                 "line_index",
                 "code",
                 "day_date"], inplace=True)
df_regression.shape

# We only need to train on delayed trains on line
df_regression = df_regression[df_regression['arrival_delay'] > 0]
df_regression.reset_index(inplace=True, drop=True)


X_train, X_test, y_train, y_test = train_test_split(df_regression.loc[:, df_regression.columns != "arrival_delay"].values , df_regression["arrival_delay"].values , test_size=0.20, random_state=10)
# To avoid excessive memory usage we delete the dataset
del df_regression

print(f"X_train: ",{X_train.shape})
print(f"X_test: ",{X_test.shape})
print(f"y_train: ",{y_train.shape})
print(f"y_test: ",{y_test.shape})

# We initialize a dataset to easily call for see statistics
statistics = pd.DataFrame(columns=['regressor','mse','rmse','mae','r2'])


# Regression

## Linear Regression

model = LinearRegression(n_jobs=-1)
model.fit(X_train, y_train)
joblib.dump(model, "./models/regressions/LinearRegression.joblib")
# model = joblib.load("./models/regressions/LinearRegression.joblib")
y_pred = model.predict(X_test)
statistics = helper.regressor_metrics(y_test, y_pred, "Linear", statistics)

del model
del y_pred

## Decision Tree Regressor

model = DecisionTreeRegressor()
model.fit(X_train, y_train)
joblib.dump(model, "./models/regressions/DecisionTreeRegressor.joblib")
#model = joblib.load("./models/regressions/DecisionTreeRegressor.joblib")
y_pred = model.predict(X_test)
statistics = helper.regressor_metrics(y_test, y_pred, "DecisionTree", statistics)
del model
del y_pred

## Random Forest

model = RandomForestRegressor(n_jobs=-1)
model.fit(X_train, y_train)
joblib.dump(model, "./models/regressions/RandomForestRegressor.joblib")
#model = joblib.load("./models/regressions/RandomForestRegressor.joblib")
y_pred = model.predict(X_test)
statistics = helper.regressor_metrics(y_test, y_pred, "RandomForest", statistics)
del model
del y_pred

## Extra Trees Regressor

model = ExtraTreesRegressor(n_jobs=-1)
model.fit(X_train, y_train)
joblib.dump(model, "./models/regressions/ExtraTreesRegressor.joblib")
#model = joblib.load("./models/regressions/ExtraTreesRegressor.joblib")
y_pred = model.predict(X_test)
statistics = helper.regressor_metrics(y_test, y_pred, "ExtraTrees", statistics)
copy_X = y_pred.copy()
del model
del y_pred

## Gradient Boosting Regressor

model = GradientBoostingRegressor()
model.fit(X_train, y_train)
joblib.dump(model, "./models/regressions/GradientBoostingRegressor.joblib")
#model = joblib.load("./models/regressions/GradientBoostingRegressor.joblib")
y_pred = model.predict(X_test)
statistics = helper.regressor_metrics(y_test, y_pred, "GradientBoosting", statistics)
del model
del y_pred

## Bagging Regressor


model = BaggingRegressor()
model.fit(X_train, y_train)
joblib.dump(model, "./models/regressions/BaggingRegressor.joblib")
#model = joblib.load("./models/regressions/BaggingRegressor.joblib")
y_pred = model.predict(X_test)
statistics = helper.regressor_metrics(y_test, y_pred, "Bagging", statistics)
del model
del y_pred

## AdaBoostRegressor


model = AdaBoostRegressor()
model.fit(X_train, y_train)
joblib.dump(model, "./models/regressions/AdaBoostRegressor.joblib")
#model = joblib.load("./models/regressions/AdaBoostRegressor.joblib")
y_pred = model.predict(X_test)
statistics = helper.regressor_metrics(y_test, y_pred, "AdaBoost", statistics)
del model
del y_pred

# Result

statistics.head(10)
